requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.140
botocore==1.34.140
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cobble==0.1.4
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
doc2docx==0.2.4
docx2txt==0.8
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
gliner==0.2.7
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gradio==4.37.2
gradio_client==1.0.2
gradio_pdf==0.0.11
greenlet==3.0.3
grpcio==1.64.1
guidance==0.1.15
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.84
llama-cloud==0.0.6
llama-index==0.10.53
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.53.post1
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-guidance==0.1.2
llama-index-program-openai==0.1.6
llama-index-question-gen-guidance==0.1.2
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.29
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.6
lxml==5.2.2
mammoth==1.8.0
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.0.1
mplcursors==0.5.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nanoid==2.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
openai==1.35.15
opencv-python==4.10.0.84
ordered-set==4.1.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdftext==0.3.10
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydot==2.0.0
pydub==0.25.1
pydyf==0.10.0
pyformlang==1.0.10
Pygments==2.18.0
PyMuPDF==1.24.7
pymupdf4llm==0.0.8
PyMuPDFb==1.24.6
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.15.0
pysqlite3==0.5.3
python-dateutil==2.9.0.post0
python-doctr @ git+https://github.com/mindee/doctr.git@1cea7d8f7db2c3cf7365be3c7457b7fe67f82ab9
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruff==0.5.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
striprtf==0.0.26
surya-ocr==0.4.14
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
texify==0.1.10
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
zopfli==0.2.3
